% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_trn}
\alias{sim_trn}
\title{Simulation of transcriptional regulatory networks (TRN)}
\usage{
sim_trn(n = 10, nreg = 3, ngene = 3, seed = NULL, reg_info = NULL,
  mod_info = NULL)
}
\arguments{
\item{n}{number of samples (data points) to generate.}

\item{nreg}{number of regulators.}

\item{ngene}{number of genes.}

\item{seed}{seed for random number generator.}

\item{reg_info}{data.frame or matrix with information about regulator's mean
and variance expression levels.}

\item{mod_info}{matrix with information about TRN}
}
\value{
a list with network information.
}
\description{
Simulation of transcriptional regulatory networks (TRN)
}

