% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{simtrnUI}
\alias{simtrnUI}
\title{simtrnUI}
\usage{
simtrnUI()
}
\description{
Runs a shiny application in a web browser that enables to interactively
experiment with TRMs and their effect in the observed gene-wise correlations.
}

