% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{trmsimUI}
\alias{trmsimUI}
\title{trmsimUI}
\usage{
trmsimUI()
}
\description{
Runs a shiny application in a web browser that enables to interactively
experiment with TRMs and their effect in the observed gene-wise correlations.
}

